requirements.txt
streamlit
transformers
torch
python-dotenv
accelerate
sentencepiece
